pyproject.toml
[project]
name = "exchange-knowledge-agent"
version = "0.1.0"
description = "Microsoft Exchange Knowledge AI Agent for processing and storing email information"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "exchangelib"
]
authors = ["nuuner <domen.iglic@proton.me>"]